track my new kernel for falcon
had to rebase and i made a new git comunity to keep my kernels and other projects in
cm updated and fixed mem leaks but changed commit history so i had to rebase
this kernel is alwase buildable. EOS tracks it too i do all tsting on a diffrent
branch.

Change-Id: I4b73e258438daec0c9a56500c1ca131286b9a30b